+++ packages/design-system/components/native/displays/index.ts
@@ -1,2 +1,3 @@
 export { default as FeedbackScreen } from './FeedbackScreen'
 export { default as LoadingScreen } from './LoadingScreen'
+export type * from './LoadingScreen/types'